Refactor etcd transactions
