+++ backend/ringv2/ringv2.go
@@ -13,7 +13,7 @@ import (
 	"github.com/coreos/etcd/mvcc/mvccpb"
 	cron "github.com/robfig/cron/v3"
 	"github.com/sensu/sensu-go/backend/store"
-	"github.com/sensu/sensu-go/backend/store/etcd"
+	"github.com/sensu/sensu-go/backend/store/etcd/kvc"
 	"github.com/sensu/sensu-go/util/retry"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/time/rate"
@@ -170,12 +170,12 @@ func New(client *clientv3.Client, storePath string) *Ring {
 // IsEmpty returns true if there are no items in the ring.
 func (r *Ring) IsEmpty(ctx context.Context) (bool, error) {
 	var resp *clientv3.GetResponse
-	err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		resp, err = r.client.Get(ctx, r.itemPrefix,
 			clientv3.WithKeysOnly(),
 			clientv3.WithPrefix(),
 			clientv3.WithLimit(1))
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return false, err
@@ -200,9 +200,9 @@ func (r *Ring) Add(ctx context.Context, value string, keepalive int64) (rerr err
 	itemKey := path.Join(r.itemPrefix, value)
 
 	var getresp *clientv3.GetResponse
-	err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		getresp, err = r.client.Get(ctx, itemKey)
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return fmt.Errorf("couldn't add %q to ring: %s", value, err)
@@ -230,9 +230,9 @@ func (r *Ring) Add(ctx context.Context, value string, keepalive int64) (rerr err
 NEWLEASE:
 
 	var lease *clientv3.LeaseGrantResponse
-	err = etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err = kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		lease, err = r.client.Grant(ctx, keepalive)
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return fmt.Errorf("couldn't add %q to ring: %s", value, err)
@@ -243,9 +243,9 @@ NEWLEASE:
 		}
 	}()
 
-	err = etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err = kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		_, err = r.client.Put(ctx, itemKey, "", clientv3.WithLease(lease.ID))
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return fmt.Errorf("couldn't add %q to ring: %s", value, err)
@@ -270,9 +270,9 @@ func (r *Ring) Remove(ctx context.Context, value string) error {
 	// Try to get the item and revoke its lease if found
 	var getresp *clientv3.GetResponse
 	key := path.Join(r.itemPrefix, value)
-	err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		getresp, err = r.client.Get(ctx, key)
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return err
@@ -285,9 +285,9 @@ func (r *Ring) Remove(ctx context.Context, value string) error {
 			_, _ = r.client.Revoke(ctx, leaseID)
 		}
 	}
-	return etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	return kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		_, err = r.client.Delete(ctx, path.Join(r.itemPrefix, value))
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 }
 
@@ -349,9 +349,9 @@ func (w *watcher) hasTrigger(ctx context.Context) (bool, string, error) {
 	getTrigger := clientv3.OpGet(w.triggerKey())
 	getFirst := clientv3.OpGet(w.ring.itemPrefix, clientv3.WithPrefix(), clientv3.WithLimit(1))
 	var resp *clientv3.TxnResponse
-	err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		resp, err = w.ring.client.Txn(ctx).Then(getTrigger, getFirst).Commit()
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return false, "", err
@@ -403,7 +403,7 @@ func (w *watcher) ensureActiveTrigger(ctx context.Context) error {
 
 		resp, err := w.ring.client.Txn(ctx).If(triggerCmp).Then(triggerOp).Commit()
 		if err != nil {
-			return etcd.RetryRequest(retry, err)
+			return kvc.RetryRequest(retry, err)
 		}
 		if !resp.Succeeded {
 			_, _ = w.ring.client.Revoke(ctx, lease.ID)
@@ -496,9 +496,9 @@ func (r *Ring) nextInRing(ctx context.Context, prevKv *mvccpb.KeyValue, n int64)
 		opts = append(opts, clientv3.WithRange(end))
 	}
 	var resp *clientv3.GetResponse
-	err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		resp, err = r.client.Get(ctx, key, opts...)
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return nil, fmt.Errorf("couldn't get next item(s) in ring: %s", err)
@@ -514,9 +514,9 @@ func (r *Ring) nextInRing(ctx context.Context, prevKv *mvccpb.KeyValue, n int64)
 	if int64(len(result)) < n {
 		m := n - int64(len(result))
 		var resp *clientv3.GetResponse
-		err := etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+		err := kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 			resp, err = r.client.Get(ctx, r.itemPrefix, clientv3.WithPrefix(), clientv3.WithLimit(m))
-			return etcd.RetryRequest(n, err)
+			return kvc.RetryRequest(n, err)
 		})
 		if err != nil {
 			return nil, fmt.Errorf("couldn't get next item(s) in ring: %s", err)
@@ -572,9 +572,9 @@ func (w *watcher) advanceRing(ctx context.Context, prevKv *mvccpb.KeyValue) ([]*
 	triggerCmp := clientv3.Compare(clientv3.Version(w.triggerKey()), "=", 0)
 
 	var resp *clientv3.TxnResponse
-	err = etcd.Backoff(ctx).Retry(func(n int) (done bool, err error) {
+	err = kvc.Backoff(ctx).Retry(func(n int) (done bool, err error) {
 		resp, err = w.ring.client.Txn(ctx).If(triggerCmp).Then(triggerOp).Commit()
-		return etcd.RetryRequest(n, err)
+		return kvc.RetryRequest(n, err)
 	})
 	if err != nil {
 		return nil, fmt.Errorf("couldn't advance ring: %s", err)
