+++ backend/store/etcd/kvc/logger.go
@@ -1,0 +1,7 @@
+package kvc
+
+import "github.com/sirupsen/logrus"
+
+var logger = logrus.WithFields(logrus.Fields{
+	"component": "etcd.kvc",
+})
